+++ data/options.js
@@ -1,31 +1,17 @@
-/* Source code from: https://developer.chrome.com/extensions/options */
 'use strict';
 
 let backgroundPage = chrome.extension.getBackgroundPage();
 let appGlobal = backgroundPage.appGlobal;
+let options_default = appGlobal.options_default;
 let getPreferences = appGlobal.getPreferences;
 let getBooleanFromVar = appGlobal.getBooleanFromVar;
 let _ = appGlobal._;
 let translateNodes = appGlobal.translateNodes;
 let translateNodes_title = appGlobal.translateNodes_title;
-let getValueFromNode = appGlobal.getValueFromNode;
-
-// Saves options to chrome.storage
-function save_options() {
-	let preferenceInput_nodes = document.querySelectorAll(".preferenceInput");
-	
-	for(let i in preferenceInput_nodes){
-		let node = preferenceInput_nodes[i];
-		if(typeof node.id == "string"){
-			port_options.sendData("setting_Update", {settingName: node.id, settingValue: getValueFromNode(node)});
-		}
-	}
-	var status = document.querySelector('#status');
-	status.textContent = _("options_saved");
-	setTimeout(function() {
-		status.textContent = '';
-	}, 1250);
+let settingNode_onChange = function(){
+	appGlobal.settingNode_onChange(event, this, port_options);
 }
+let getValueFromNode = appGlobal.getValueFromNode;
 
 function sendDataToMain(portName){
 	this.port = chrome.runtime.connect({name: portName});
@@ -42,82 +28,362 @@ port.onDisconnect.addListener(function(port) {
 	console.info(`Port disconnected: ${port.name}`);
 });
 
-let dailymotion_import_button = document.querySelector("button#dailymotion_import");
-dailymotion_import_button.addEventListener("click", function(){
-	port_options.sendData("importStreams","dailymotion");
-});
-
-let hitbox_import_button = document.querySelector("button#hitbox_import");
-hitbox_import_button.addEventListener("click", function(){
-	port_options.sendData("importStreams","hitbox");
-});
-
-let twitch_import_button = document.querySelector("button#twitch_import");
-twitch_import_button.addEventListener("click", function(){
-	port_options.sendData("importStreams","twitch");
-});
-
-let beam_import_button = document.querySelector("button#beam_import");
-beam_import_button.addEventListener("click", function(){
-	port_options.sendData("importStreams","beam");
-});
-
-let pref_nodes = {
-	dailymotion_keys_list: document.querySelector('#dailymotion_keys_list'),
-	hitbox_keys_list: document.querySelector('#hitbox_keys_list'),
-	twitch_keys_list: document.querySelector('#twitch_keys_list'),
-	beam_keys_list: document.querySelector('#beam_keys_list'),
-	
-	dailymotion_user_id: document.querySelector('#dailymotion_user_id'),
-	hitbox_user_id: document.querySelector('#hitbox_user_id'),
-	twitch_user_id: document.querySelector('#twitch_user_id'),
-	beam_user_id: document.querySelector('#beam_user_id'),
-
-	check_delay:					document.querySelector('#check_delay'),
-	notification_type:				document.querySelector('#notification_type'),
-	notify_online:					document.querySelector('#notify_online'),
-	notify_offline:					document.querySelector('#notify_offline'),
-	show_offline_in_panel:			document.querySelector('#show_offline_in_panel'),
-	confirm_addStreamFromPanel:		document.querySelector('#confirm_addStreamFromPanel'),
-	confirm_deleteStreamFromPanel:	document.querySelector('#confirm_deleteStreamFromPanel'),
-	panel_theme:					document.querySelector('#panel_theme'),
-	background_color: 				document.querySelector('#background_color'),
-	
-	livestreamer_cmd_to_clipboard: 	document.querySelector('#livestreamer_cmd_to_clipboard'),
-	livestreamer_cmd_quality: 		document.querySelector('#livestreamer_cmd_quality')
+let options = {
+	"dailymotion": {
+		type: "group",
+		prefs: {
+			"dailymotion_keys_list": {
+				"title": "Dailymotion keys to notify",
+				"description": "Stream list in a comma separated list.",
+				"type": "string",
+			},
+			"dailymotion_user_id": {
+				"title": "Your Dailymotion id",
+				"description": "Enter your Dailymotion id to be able to import the stream(s) you follow.",
+				"type": "string",
+			},
+			"dailymotion_import": {
+				"title": "Import the Dailymotion stream(s) you follow",
+				"label": "Import",
+				"type": "control"
+			}
+		}
+	},
+	"hitbox": {
+		type: "group",
+		prefs: {
+			"hitbox_keys_list": {
+				"title": "Hitbox keys to notify",
+				"description": "Stream list in a comma separated list.",
+				"type": "string",
+			},
+			"hitbox_user_id": {
+				"title": "Your Hitbox id",
+				"description": "Enter your Hitbox id to be able to import the stream(s) you follow.",
+				"type": "string",
+			},
+			"hitbox_import": {
+				"title": "Import the Hitbox stream(s) you follow",
+				"label": "Import",
+				"type": "control"
+			}
+		}
+	},
+	"twitch": {
+		type: "group",
+		prefs: {
+			"twitch_keys_list": {
+				"title": "Twitch keys to notify",
+				"description": "Stream list in a comma separated list.",
+				"type": "string",
+			},
+			"twitch_user_id": {
+				"title": "Your Twitch id",
+				"description": "Enter your Twitch id to be able to import the stream(s) you follow.",
+				"type": "string",
+			},
+			"twitch_import": {
+				"title": "Import the Twitch stream(s) you follow.",
+				"label": "Import",
+				"type": "control"
+			}
+		}
+	},
+	"beam": {
+		type: "group",
+		prefs: {
+			"beam_keys_list": {
+				"title": "Beam keys to notify",
+				"description": "Stream list in a comma separated list.",
+				"type": "string",
+			},
+			"beam_user_id": {
+				"title": "Your Beam id",
+				"description": "Enter your Beam id to be able to import the stream(s) you follow.",
+				"type": "string",
+			},
+			"beam_import": {
+				"title": "Import the Beam stream(s) you follow.",
+				"label": "Import",
+				"type": "control"
+			}
+		}
+	},
+	"notification_type": {
+		"tittle": "Notification type",
+		"description": "",
+		"type": "menulist",
+		"options": [
+				{
+					"value": "web",
+					"label": "Web"
+				},
+				{
+					"value": "chrome_api",
+					"label": "Chrome API"
+				}
+			]
+	},
+	"check_delay": {
+		"title": "Streams status delay",
+		"description": "Delay between checks, in minute",
+		"type": "integer",
+		"value": 5
+	},
+	"notify": {
+		type: "group",
+		prefs: {
+			"notify_online": {
+				"title": "Show a notification when a stream start",
+				"description": "Notification when checked",
+				"type": "bool",
+			},
+			"notify_offline": {
+				"title": "Show a notification when a stream finish",
+				"description": "Notification when checked",
+				"type": "bool",
+			}
+		}
+	},
+	"showInPanel": {
+		type: "group",
+		prefs: {
+			"group_streams_by_websites": {
+				"title": "Group streams by website",
+				"description": "Grouped when checked",
+				"type": "bool",
+			},
+			"show_offline_in_panel": {
+				"title": "Show offline streams in the panel",
+				"description": "Shown when checked",
+				"type": "bool",
+			}
+		}
+	},
+	"confirmAddDelete": {
+		type: "group",
+		prefs: {
+			"confirm_addStreamFromPanel": {
+				"title": "Confirmation to add streams",
+				"description": "Show a notification to confirm when adding a stream of config (from panel)",
+				"type": "bool",
+			},
+			"confirm_deleteStreamFromPanel": {
+				"title": "Confirmation to delete streams",
+				"description": "Show a notification to confirm when deleting a stream of config (from panel)",
+				"type": "bool",
+			}
+		}
+	},
+	"theme": {
+		type: "group",
+		prefs: {
+			"panel_theme": {
+				"title": "Panel theme",
+				"description": "Choose the panel of the panel",
+				"type": "menulist",
+				"options": [
+						{
+							"value": "dark",
+							"label": "Dark"
+						},
+						{
+							"value": "light",
+							"label": "Light"
+						}
+					]
+			},
+			"background_color": {
+				"title": "Panel background color",
+				"description": "Choose background color",
+				"type": "color",
+			}
+		}
+	},
+	"livestreamer": {
+		type: "group",
+		prefs: {
+			"livestreamer_cmd_to_clipboard": {
+				"title": "Copy Livestreamer command on stream click.",
+				"description": "Check to activate",
+				"type": "bool",
+			},
+			"livestreamer_cmd_quality": {
+				"title": "Livestreamer quality",
+				"description": "More information on Livestreamer page.",
+				"type": "string",
+			}
+		}
+	}
 }
-function restore_options() {
-	pref_nodes.dailymotion_keys_list.value =			getPreferences("dailymotion_keys_list");
-	pref_nodes.hitbox_keys_list.value =					getPreferences("hitbox_keys_list");
-	pref_nodes.twitch_keys_list.value =					getPreferences("twitch_keys_list");
-	pref_nodes.beam_keys_list.value =					getPreferences("beam_keys_list");
+
+function loadPreferences(){
+	let container = document.querySelector("section#preferences");
 	
-	pref_nodes.dailymotion_user_id.value =				getPreferences("dailymotion_user_id");
-	pref_nodes.hitbox_user_id.value =					getPreferences("hitbox_user_id");
-	pref_nodes.twitch_user_id.value =					getPreferences("twitch_user_id");
-	pref_nodes.beam_user_id.value =						getPreferences("beam_user_id");
+	for(let id in options){
+		let option = options[id];
+		if(typeof option.type == "undefined"){
+			continue;
+		}
+		if(option.type == "group"){
+			newPreferenceGroup(container, id, option.prefs);
+		} else {
+			newPreferenceNode(container, id, option);
+		}
+	}
+}
+function newPreferenceGroup(parent, groupId, prefs){
+	let groupNode = document.createElement("p");
+	groupNode.id = groupId;
+	groupNode.className = "pref_group";
+	if(groupId == "dailymotion" || groupId == "hitbox" || groupId == "twitch" || groupId == "beam"){
+		groupNode.className += " website_pref"
+	}
+	for(let id in prefs){
+		if(typeof prefs[id].type == "undefined"){
+			continue;
+		}
+		newPreferenceNode(groupNode, id, prefs[id]);
+	}
 	
-	pref_nodes.check_delay.value =						parseInt(getPreferences("check_delay"));
-	pref_nodes.notification_type.value = 				getPreferences("notification_type");
-	pref_nodes.notify_online.checked =					getPreferences("notify_online");
-	pref_nodes.notify_offline.checked =					getPreferences("notify_offline");
-	pref_nodes.show_offline_in_panel.checked =			getPreferences("show_offline_in_panel");
-	pref_nodes.confirm_addStreamFromPanel.checked =		getPreferences("confirm_addStreamFromPanel");
-	pref_nodes.confirm_deleteStreamFromPanel.checked =	getPreferences("confirm_deleteStreamFromPanel");
-	pref_nodes.panel_theme.value =						getPreferences("panel_theme");
-	pref_nodes.background_color.value =					getPreferences("background_color");
+	parent.appendChild(groupNode);
+}
+function import_onClick(){
+	let getWebsite = /^(\w+)_import$/i;
+	let website = getWebsite.exec(this.id)[1];
+	port_options.sendData("importStreams", website);
+}
+function newPreferenceNode(parent, id, prefObj){
+	let node = document.createElement("div");
+	node.className = "preferenceContainer";
 	
-	pref_nodes.livestreamer_cmd_to_clipboard.checked =	getPreferences("livestreamer_cmd_to_clipboard");
-	pref_nodes.livestreamer_cmd_quality.value =			getPreferences("livestreamer_cmd_quality");
+	let labelNode = document.createElement("label");
+	labelNode.className = "preference";
+	if(typeof prefObj.description == "string"){
+		labelNode.title = prefObj.description;
+	}
+	labelNode.htmlFor = id;
+	labelNode.setAttribute("data-translate-title",`${id}_description`)
+	
+	let title = document.createElement("span");
+	title.id = `${id}_title`;
+	title.textContent = prefObj.title
+	title.setAttribute("data-translate-id",`${id}_title`)
+	labelNode.appendChild(title);
+	
+	let prefNode = null;
+	switch(prefObj.type){
+		case "string":
+			prefNode = document.createElement("input");
+			prefNode.type = "text";
+			prefNode.value = getPreferences(id);
+			break;
+		case "integer":
+			prefNode = document.createElement("input");
+			prefNode.type = "number";
+			prefNode.value = parseInt(getPreferences(id));
+			break;
+		case "bool":
+			prefNode = document.createElement("input");
+			prefNode.type = "checkbox";
+			prefNode.checked = getBooleanFromVar(getPreferences(id));
+			break;
+		case "color":
+			prefNode = document.createElement("input");
+			prefNode.type = "color";
+			prefNode.value = getPreferences(id);
+			break;
+		case "control":
+			prefNode = document.createElement("button");
+			prefNode.textContent = prefObj.label;
+			break;
+		case "menulist":
+			prefNode = document.createElement("select");
+			prefNode.size = 2;
+			for(let o in prefObj.options){
+				let option = prefObj.options[o];
+				
+				let optionNode = document.createElement("option");
+				optionNode.text = option.label;
+				optionNode.value = option.value;
+				optionNode.setAttribute("data-translate-id",`${id}_${option.value}`);
+				
+				prefNode.add(optionNode);
+			}
+			prefNode.value = getPreferences(id);
+			break;
+	}
+	prefNode.id = id;
+	if(prefObj.type != "control"){
+		prefNode.className = "preferenceInput";
+	}
+	if(id.indexOf("_keys_list") != -1){
+		node.className += " flex_input_text";
+	}
+	prefNode.setAttribute("data-setting-type", prefObj.type);
+	
+	if(prefObj.type != "menulist"){
+		prefNode.setAttribute("data-translate-id", id);
+	}
+	
+	node.appendChild(labelNode);
+	node.appendChild(prefNode);
+	parent.appendChild(node);
+	
+	switch(prefObj.type){
+		case "string":
+			prefNode.addEventListener("input", settingNode_onChange, false);
+			break;
+		case "integer":
+		case "bool":
+		case "color":
+		case "menulist":
+			prefNode.addEventListener("change", settingNode_onChange, false);
+			break;
+		case "control":
+			if(id.indexOf("_import") != -1){
+				prefNode.addEventListener("click", import_onClick, false);
+			}
+			break;
+	}
+}
+loadPreferences();
+
+function refresh_options(){
+	let prefNodes = document.querySelectorAll(".preferenceInput");
+	for(let i in prefNodes){
+		if(typeof prefNodes[i].tagName == "undefined"){
+			continue;
+		}
+		
+		let prefNode = prefNodes[i];
+		let settingType = prefNode.dataset.settingType;
+		let prefId = prefNode.id;
+		let prefValue = getPreferences(prefId);
+		switch(settingType){
+			case "string":
+			case "color":
+			case "menulist":
+				prefNode.value = prefValue;
+				break;
+			case "integer":
+				prefNode.value = parseInt(prefValue);
+				break;
+			case "bool":
+				prefNode.checked = getBooleanFromVar(prefValue);
+				break;
+			case "control":
+				// Nothing to update, no value
+				break;
+		}
+	}
 }
 
 function init(){
 	translateNodes(document);
 	translateNodes_title(document);
-	restore_options();
 }
-document.addEventListener('DOMContentLoaded',				init);
-document.querySelector('#save').addEventListener('click',	save_options);
+document.addEventListener('DOMContentLoaded',		init);
 
 let port_mainscript = null
 chrome.runtime.onConnect.addListener(function(_port) {
@@ -134,7 +400,7 @@ chrome.runtime.onConnect.addListener(function(_port) {
 		
 		switch(id){
 			case "refreshOptions":
-				restore_options();
+				refresh_options();
 				break;
 		}
 	});
@@ -144,113 +410,93 @@ chrome.runtime.onConnect.addListener(function(_port) {
 	});
 });
 
+// Saves options to chrome.storage
 // Storage area compatibility - Might not be useful for now, but Firefox Webextension doesn't seems to plan using the sync storage area
 let storage = (typeof chrome.storage.sync == "object")? chrome.storage.sync : chrome.storage.local;
 
 // Save states using in chrome.storage.
 function saveOptionsInSync(){
-	let dailymotion_keys_list = pref_nodes.dailymotion_keys_list.value;
-	let hitbox_keys_list = pref_nodes.hitbox_keys_list.value;
-	let twitch_keys_list = pref_nodes.twitch_keys_list.value;
-	let beam_keys_list = pref_nodes.beam_keys_list.value;
+	let settingsDataToSync = {};
 	
-	let dailymotion_user_id = pref_nodes.dailymotion_user_id.value;
-	let hitbox_user_id = pref_nodes.hitbox_user_id.value;
-	let twitch_user_id = pref_nodes.twitch_user_id.value;
-	let beam_user_id = pref_nodes.beam_user_id.value;
+	let prefNodes = document.querySelectorAll(".preferenceInput");
+	for(let i in prefNodes){
+		let prefNode = prefNodes[i]
+		if(typeof prefNode.tagName == "undefined"){
+			continue;
+		}
+		let settingType = prefNode.dataset.settingType;
+		let prefId = prefNode.id;
+		switch(settingType){
+			case "string":
+			case "color":
+			case "menulist":
+				settingsDataToSync[prefId] = prefNode.value;
+				break;
+			case "integer":
+				settingsDataToSync[prefId] = parseInt(prefNode.value);
+				break;
+			case "bool":
+				settingsDataToSync[prefId] = getBooleanFromVar(prefNode.checked);
+				break;
+			case "control":
+				// Nothing to update, no value
+				break;
+		}
+	}
 	
-	let check_delay =						parseInt(pref_nodes.check_delay.value);
-	let notification_type =					pref_nodes.notification_type.value;
-	let notify_online =						pref_nodes.notify_online.checked;
-	let notify_offline =					pref_nodes.notify_offline.checked;
-	let show_offline_in_panel =				pref_nodes.show_offline_in_panel.checked;
-	let confirm_addStreamFromPanel =		pref_nodes.confirm_addStreamFromPanel.checked;
-	let confirm_deleteStreamFromPanel =		pref_nodes.confirm_deleteStreamFromPanel.checked;
-	let panel_theme =						pref_nodes.panel_theme.value;
-	let background_color = 					pref_nodes.background_color.value;
-	let livestreamer_cmd_to_clipboard = 	pref_nodes.livestreamer_cmd_to_clipboard.checked;
-	let livestreamer_cmd_quality = 			pref_nodes.livestreamer_cmd_quality.value;
-	storage.set({
-		dailymotion_keys_list: dailymotion_keys_list,
-		hitbox_keys_list: hitbox_keys_list,
-		twitch_keys_list: twitch_keys_list,
-		beam_keys_list: beam_keys_list,
-		
-		dailymotion_user_id: dailymotion_user_id,
-		hitbox_user_id: hitbox_user_id,
-		twitch_user_id: twitch_user_id,
-		beam_user_id: beam_user_id,
-		
-		check_delay: check_delay,
-		notification_type: notification_type,
-		notify_online: notify_online,
-		notify_offline: notify_offline,
-		show_offline_in_panel: show_offline_in_panel,
-		confirm_addStreamFromPanel: confirm_addStreamFromPanel,
-		confirm_deleteStreamFromPanel: confirm_deleteStreamFromPanel,
-		panel_theme: panel_theme,
-		background_color: background_color,
-		livestreamer_cmd_to_clipboard: livestreamer_cmd_to_clipboard,
-		livestreamer_cmd_quality: livestreamer_cmd_quality
-	}, function() {
+	storage.set(settingsDataToSync, function() {
 		// Update status to let user know options were saved.
 		var status = document.getElementById('status');
 		status.textContent = _("options_saved_sync");
+		if(typeof chrome.runtime.lastError != "undefined"){
+			console.warn(chrome.runtime.lastError);
+		}
 		setTimeout(function() {
 			status.textContent = '';
-		}, 1250);
+		}, 2000);
 	});
 }
-// Restores states using the preferences
-// stored in chrome.storage.
+// Restores states using the preferences stored in chrome.storage.
 function restaureOptionsFromSync(){
 	// Default values
-	storage.get({
-		dailymotion_keys_list: "",
-		hitbox_keys_list: "",
-		twitch_keys_list: "",
-		beam_keys_list: "",
+	storage.get(options_default, function(items) {
+		if(typeof chrome.runtime.lastError != "undefined"){
+			console.warn(chrome.runtime.lastError);
+		}
 		
-		dailymotion_user_id: "",
-		hitbox_user_id: "",
-		twitch_user_id: "",
-		beam_user_id: "",
-		
-		check_delay: 5,
-		notification_type: "web",
-		notify_online: true,
-		notify_offline: false,
-		show_offline_in_panel: false,
-		confirm_addStreamFromPanel: false,
-		confirm_deleteStreamFromPanel: true,
-		panel_theme: "dark",
-		background_color: "#000000",
-		
-		livestreamer_cmd_to_clipboard: false,
-		livestreamer_cmd_quality: "best"
-	}, function(items) {
-		pref_nodes.dailymotion_keys_list.value =				items.dailymotion_keys_list;
-		pref_nodes.hitbox_keys_list.value =						items.hitbox_keys_list;
-		pref_nodes.twitch_keys_list.value =						items.twitch_keys_list;
-		pref_nodes.beam_keys_list.value =						items.beam_keys_list;
-		
-		pref_nodes.dailymotion_user_id.value =					items.dailymotion_user_id;;
-		pref_nodes.hitbox_user_id.value =						items.hitbox_user_id;;
-		pref_nodes.twitch_user_id.value =						items.twitch_user_id;
-		pref_nodes.beam_user_id.value =							items.beam_user_id;
-		
-		pref_nodes.check_delay.value =							parseInt(items.check_delay);
-		pref_nodes.notification_type.value =					items.notification_type;
-		pref_nodes.notify_online.checked =						getBooleanFromVar(items.notify_online);
-		pref_nodes.notify_offline.checked =						getBooleanFromVar(items.notify_offline);
-		pref_nodes.show_offline_in_panel.checked =				getBooleanFromVar(items.show_offline_in_panel);
-		pref_nodes.confirm_addStreamFromPanel.checked =			getBooleanFromVar(items.confirm_addStreamFromPanel);
-		pref_nodes.confirm_deleteStreamFromPanel.checked =		getBooleanFromVar(items.confirm_deleteStreamFromPanel);
-		pref_nodes.panel_theme.value =							items.panel_theme;
-		pref_nodes.background_color.value =						items.background_color;
-		
-		pref_nodes.livestreamer_cmd_to_clipboard.checked =		items.livestreamer_cmd_to_clipboard;
-		pref_nodes.livestreamer_cmd_quality.value =				items.livestreamer_cmd_quality;
+		for(let id in items){
+			let prefNode = document.querySelector(`#${id}`);
+			
+			if(prefNode == null){
+				console.group();
+				console.warn(`Sync restaure: the node of the preference ${id} doesn't not exist in the option page (value: ${items[id]})`);
+				console.info("Maybe it's just a setting without sync (addon version, for exemple)")
+				console.groupEnd();
+				return;
+			}
+			
+			let settingType = prefNode.dataset.settingType;
+			let prefId = prefNode.id;
+			switch(settingType){
+				case "string":
+				case "color":
+				case "menulist":
+					prefNode.value = items[id];
+					break;
+				case "integer":
+					prefNode.value = parseInt(items[id]);
+					break;
+				case "bool":
+					prefNode.checked = getBooleanFromVar(items[id]);
+					break;
+				case "control":
+					// Nothing to update, no value
+					break;
+			}
+			
+			/*				Save the settings received from sync				*/
+			appGlobal.settingNode_onChange({type: "change"}, prefNode, port_options);
+		}
 	});
 }